requirements.txt
ScrapyElasticSearch==0.8.9

